+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/exporters/settings/ShapeExportSettings.java
@@ -48,7 +48,7 @@ public class ShapeExportSettings {
             case SWF:
                 return ".swf";
             default:
-                throw new Error("Unsupported morphshape export mode: " + mode);
+                throw new Error("Unsupported shape export mode: " + mode);
         }
     }
 }